Avatar refactor * split to many files * iconTypes -> ICON_TYPE * iconColors -> ICON_COLOR
